Moved ProductHolderBase to its own header and source file
By removing the base class into its own header file we reduce
coupling to the different implementations of the inheriting classes.
